backend/requirements.txt
mysql-connector-python
python-geoip
python-geoip-geolite2
python-dateutil
fastapi
sqlalchemy
uvicorn
